==> contexts/02_Backtranslate_English.ipynb <==
# 02 English Back Translate (Translation done on 11/07/2023)

- Input: `02_backtranslate_english_input.csv`
- Output: `02_backtranslate_english.csv`

This file is for sanity checking and processing the translated files. You should already have taken the `01_prepare_dataset.csv` and use Google Sheets to translate Thai Questions and back translate it, and save it into `02_backtranslate_english_input.csv`. This file will check if the back translation format is correct.
- English Translation should be named `en_aug`
- Back Translation should be named `th_aug`

import pandas as pd
from utils import clean_text

backtranslated = pd.read_csv("data/02_backtranslate_english_input.csv")
backtranslated

## Race Condition Handling

original = pd.read_csv("data/01_prepare_dataset.csv")
original.isna().sum()

For some weird reason, Google Sheets dropped some "context" from the rows. This could be due to some weird unicode handling in Google Sheets.

Regardless, we can rejoin the context by using the `id` and `context` from the first file:

# Sanity check for missing values
backtranslated.isnull().sum()

# Drop #VALUE! rows
backtranslated = backtranslated[backtranslated["en_aug"] != "#VALUE!"]
backtranslated

# We can see that there are 34 missing values in the `context` column. We can fill it up with the original context from the first file by removing and merging.
backtranslated = original.merge(backtranslated, on="context", how="left").dropna()
backtranslated

assert backtranslated.isnull().sum().sum() == 0

## Clean Up Back Translation

# To ensure that the back translation questions are normalized / cleaned, we will use the `clean_text` function from `utils.py` to clean up the back translation.
backtranslated["en_aug"] = backtranslated["en_aug"].apply(lambda x: clean_text(x, is_question=True))
backtranslated["th_aug"] = backtranslated["th_aug"].apply(lambda x: clean_text(x, is_question=True))

backtranslated

## Save Backtranslated Dataset

backtranslated.to_csv("data/02_backtranslate_english.csv", index=False)

backtranslated[["context", "en_aug", "th_aug"]].drop_duplicates().to_csv("data/03_augment_llm_input.csv", index=False)

backtranslated.iloc[1000]["en_aug"]
