==> contexts/04_Paraphrase_QCPG.ipynb <==
# 04 QCPG Paraphrase
- Input: `03_augment_llm.csv`
- Output: `04_qcpg_paraphrase.csv`

import gc
import platform
import pandas as pd

import torch

from datasets import Dataset
from transformers import pipeline
from transformers.pipelines.pt_utils import KeyDataset

from tqdm.notebook import tqdm

# If on mac default to mps, otherwise cuda, then cpu
if torch.cuda.is_available():
    device = torch.device("cuda:0")
else:
    # Check if on mac
    if platform.system() == "Darwin":
        device = torch.device("mps")
    else:
        device = torch.device("cpu")
device

## Load data

df = pd.read_csv("data/03_augment_llm.csv")
df

pipe = pipeline('text2text-generation', model=f'ibm/qcpg-sentences', device=device)

def num_tokens_from_string(string: str) -> int:
    """Returns the number of tokens in a text string."""
    num_tokens = int(len(string) * 0.5)
    return num_tokens

import nltk
from nltk.tokenize import sent_tokenize

def break_into_chunks(text: str, max_tokens: int = 1024) -> list:
    """Breaks the text into chunks that fit within the model's context using sentence segmentation."""
    chunks = []
    current_chunk = ""
    sentences = sent_tokenize(text)
    
    for sentence in sentences:
        temp_chunk = current_chunk + " " + sentence if current_chunk else sentence
        token_count = num_tokens_from_string(temp_chunk)
        
        if token_count <= max_tokens:
            current_chunk = temp_chunk
        else:
            chunks.append(current_chunk)
            current_chunk = sentence
    
    # Adding the last chunk if there's anything left
    if current_chunk:
        chunks.append(current_chunk)

    return chunks

def break_into_sentences(text: str) -> list:
    """Breaks the text into sentences using sentence segmentation."""
    sentences = sent_tokenize(text)
    return sentences

def run_augmentation(df, column_name, sem, lex, syn):
    df = df.copy(deep=True)
    prefix = f"COND_SEMANTIC_SIM_{int(sem*100)} COND_LEXICAL_DIV_{int(lex*100)} COND_SYNTACTIC_DIV_{int(syn*100)} "
    qcpg_augmented = []

    dataset = Dataset.from_pandas(df)

    # for out in tqdm(pipe(KeyDataset(dataset, column_name), batch_size=16), desc=f"Running QCPG", total=len(df)):
    #     qcpg_augmented.append(out[0]["generated_text"])
    
    for row in tqdm(dataset, desc=f"Running QCPG", total=len(df)):
        # chunks = break_into_chunks(row[column_name])
        chunks = break_into_sentences(row[column_name])
        finished_chunks = []
        # for chunk in chunks:
        #     out = pipe(prefix + chunk)
        #     finished_chunks.append(out[0]["generated_text"].strip())
        out = pipe([prefix + chunk for chunk in chunks])
        finished_chunks = [o["generated_text"].strip() for o in out]
        qcpg_augmented.append(" ".join(finished_chunks).strip())
        print(" ".join(finished_chunks).strip())

    gc.collect()
    
    return qcpg_augmented

settings_values = [0.2, 0.5, 0.8] 
columns = ["en_aug", "en_llm_gec_aug"]
for column in tqdm(columns, desc="Augmenting Columns"):
    for value in tqdm(settings_values, desc="Augmenting Column Values"):
        if column == "en_aug":
            target_name = f"en_qcpg_{value}_aug"
        else:
            target_name = f"en_qcpg_{value}_{'_'.join(column.split('_')[1:-1])}_aug"
        
        df[target_name] = run_augmentation(df, column, value, value, value)

# Sanity check for no empty rows
assert df.isnull().sum().sum() == 0
df

# Export and save QCPG
df.to_csv("data/04_paraphrase_qcpg.csv", index=False)

# Reformat output to make easier

import pandas as pd
import tqdm.autonotebook as tqdm

og_out = pd.read_csv('data/04_paraphrase_qcpg.csv')

selected_columns = [x for x in og_out.columns if ("aug" in x and "th_aug" != x and "en_aug" != x)]
selected_columns

import os
import zipfile

# Export each column individually
for column in selected_columns:
    if not os.path.exists(f"data/04_paraphrase_qcpg_subsets"):
        os.makedirs(f"data/04_paraphrase_qcpg_subsets")
    og_out[["context", column]].to_csv(f'data/04_paraphrase_qcpg_subsets/04_paraphrase_qcpg_{column}.csv', index=False)

# Zip them all up with highest compression
with zipfile.ZipFile('data/04_paraphrase_qcpg_subsets.zip', 'w', zipfile.ZIP_DEFLATED, compresslevel=9) as zipf:
    for column in tqdm.tqdm(selected_columns):
        zipf.write(f'data/04_paraphrase_qcpg_subsets/04_paraphrase_qcpg_{column}.csv', arcname=f'04_paraphrase_qcpg_{column}.csv')

# Merge Finished Augmentations

import pandas as pd
previous_augmentation = pd.read_csv('data/03_augment_llm.csv')
previous_augmentation

original = pd.read_csv('../questions/data/01_prepare_dataset.csv')
original["source"].value_counts()

import zipfile
import os
# Finished augmentations have same filename and is stored in 04_paraphrase_qcpg_subsets_finished
# Unzip and merge all

with zipfile.ZipFile('data/04_paraphrase_qcpg_subsets_finished.zip', 'r') as zipf:
    zipf.extractall('data/')

# Get all the files
files = [x for x in os.listdir('data/04_paraphrase_qcpg_subsets_finished/') if x.endswith('.csv')]
files

# Merge them all
merged_df = previous_augmentation.copy(deep=True) 

for file in files:
    df = pd.read_csv(f'data/04_paraphrase_qcpg_subsets_finished/{file}')

    cols_to_use = df.columns.difference(merged_df.columns).to_list()
    cols_to_use += ['context']

    merged_df = pd.merge(merged_df, df[cols_to_use], on='context', how='left')

merged_df

# Assert length is the same
assert len(merged_df) == len(previous_augmentation)

import hashlib
# Hash context column as id

def hash_string(string: str) -> str:
    """Hashes a string using sha256."""
    return hashlib.sha256(string.encode()).hexdigest()

merged_df['id'] = merged_df['context'].apply(hash_string)
merged_df

# Export results
merged_df.to_parquet('data/05_augment_thai_input.parquet', index=False)
